fix: fix admin filter, db connection leaks, and fastapi serialization
- fix TypeError in IsAdmin filter by parsing ADMINS env string to int list
- close aiohttp session and asyncpg pool on bot shutdown to prevent leaks
- convert asyncpg Record objects to dict for correct Pydantic validation
- enforce Decimal type in scraper to prevent precision loss

+++ scraper.py
@@ -1,4 +1,5 @@
 import aiohttp
+from decimal import Decimal
 
 async def get_price(symbol: str, session: aiohttp.ClientSession):
     url = f"https://api.binance.com/api/v3/ticker/price?symbol={symbol}USDT"
@@ -9,5 +10,5 @@ async def get_price(symbol: str, session: aiohttp.ClientSession):
         data = await response.json()
         return {
             "symbol": symbol,
-            "price": float(data["price"])
+            "price": Decimal(data["price"])
         }
